# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API",
    "version": "1.0",
    "description": "A simple in-memory CRUD API for managing tasks."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development server" }
  ],
  "tags": [
    { "name": "Tasks", "description": "Create, read, update and delete tasks" }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "List all tasks",
        "description": "Returns all tasks. Supports optional filtering by completion status (`done`) and case-insensitive title search (`search`); both may be combined.",
        "operationId": "listTasks",
        "parameters": [
          {
            "name": "done",
            "in": "query",
            "required": false,
            "description": "Filter by completion status.",
            "schema": { "type": "string", "enum": ["true", "false"] }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Return only tasks whose title contains this word (case-insensitive).",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "An array of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Query param 'done' must be 'true' or 'false'" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Create a new task",
        "operationId": "createTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewTask" },
              "example": { "title": "Learn Express" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Validation failed (title missing or empty)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Title is required and must be a non-empty string" }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Numeric ID of the task",
          "schema": { "type": "integer" }
        }
      ],
      "get": {
        "tags": ["Tasks"],
        "summary": "Get a single task by ID",
        "operationId": "getTask",
        "responses": {
          "200": {
            "description": "The requested task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Tasks"],
        "summary": "Update an existing task",
        "operationId": "updateTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTask" },
              "example": { "title": "Updated title", "done": true }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Done must be a boolean" }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete a task",
        "operationId": "deleteTask",
        "responses": {
          "204": { "description": "Task deleted successfully (no content)" },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get task statistics",
        "description": "Returns counts computed from the current in-memory tasks array.",
        "operationId": "getStats",
        "responses": {
          "200": {
            "description": "Aggregate task counts",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Stats" },
                "example": { "total": 3, "done": 1, "open": 2 }
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "tags": ["Tasks"],
        "summary": "Reset the store to the original example tasks",
        "description": "Replaces the in-memory tasks array with the original 3 example tasks and returns the reset list.",
        "operationId": "resetTasks",
        "responses": {
          "200": {
            "description": "The reset list of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "title": { "type": "string", "example": "Buy groceries" },
          "done": { "type": "boolean", "example": false }
        },
        "required": ["id", "title", "done"]
      },
      "NewTask": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "example": "Learn Express" }
        },
        "required": ["title"]
      },
      "UpdateTask": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "example": "Updated title" },
          "done": { "type": "boolean", "example": true }
        }
      },
      "Stats": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "example": 3 },
          "done": { "type": "integer", "example": 1 },
          "open": { "type": "integer", "example": 2 }
        },
        "required": ["total", "done", "open"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      }
    }
  }
}